Change FirewalledIDType to be a BigInteger
Usually foreign keys are big integers at the database level, so we should match the type at the ruby level

+++ lib/activerecord/firewall/firewalled_id_type.rb
@@ -1,5 +1,5 @@
 module ActiveRecord
-  class FirewalledIDType < ActiveRecord::Type::Integer
+  class FirewalledIDType < ActiveRecord::Type::BigInteger
     class FirewalledAccess < ActiveRecord::RecordNotFound; end
 
     def initialize(model, protected_type)
